requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
requests
